requirements.txt
Django==4.2.7
Pillow
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-ckeditor==6.7.0
django-taggit==4.0.0
psycopg2-binary==2.9.9
django-environ==0.11.2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Web scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
ratebeer==2.3.1
requests-cache==1.1.1
python-slugify==8.0.4
tqdm==4.66.1
fake-useragent==1.4.0
